fix the damn currency

+++ app/models/accommodation.rb
@@ -5,7 +5,7 @@ class Accommodation < ActiveRecord::Base
 
   validates :occupancy, presence: true, numericality: :integer, allow_blank: false
   validates :quantity, presence: true, numericality: :integer, allow_blank: false
-  validates :price, presence: true, numericality: :float, allow_blank: false
+  validates :price, presence: true, numericality: :decimal, allow_blank: false
 
   def self.search(search)
     if search
